% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCR.R
\name{DCR}
\alias{DCR}
\title{Disease Control Rate and Binomial Confidence interval}
\usage{
DCR(x, n)
}
\arguments{
\item{x}{a number indicating total number of subjects with Stable Disease, Partial Response, or Complete Response}

\item{n}{a number indicating total number of evaluable subjects}
}
\value{
disease control rate and 95% binomial confidence interval
}
\description{
Disease Control Rate and Binomial Confidence interval
}
\examples{
DCR(7,13)
}
\author{
Jon
}
