% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORR.R
\name{ORR}
\alias{ORR}
\title{Objective Response Rate and Binomial Confidence interval}
\usage{
ORR(x, n)
}
\arguments{
\item{x}{a number indicating total number of subjects with Partial Response or Complete Response}

\item{n}{a number indicating total number of evaluable subjects}
}
\value{
Objective Response Rate and  binomial confidence interval
}
\description{
Objective Response Rate and Binomial Confidence interval
}
\examples{
ORR(5,13)
}
\author{
Jon
}
